+++ httpdomain/sphinxcontrib/httpdomain.py
@@ -195,6 +195,8 @@ HTTP_STATUS_CODES = {
     510: 'Not Extended'
 }
 
+WEBDAV_STATUS_CODES = [207, 422, 423, 424, 507]
+
 http_sig_param_re = re.compile(r'\((?:(?P<type>[^:)]+):)?(?P<name>[\w_]+)\)',
                                re.VERBOSE)
 
@@ -453,6 +455,8 @@ class HTTPXRefStatusRole(XRefRole):
             url = 'http://tools.ietf.org/html/rfc6585#section-4'
         elif code == 449:
             url = 'http://msdn.microsoft.com/en-us/library/dd891478(v=prot.10).aspx'
+        elif code in WEBDAV_STATUS_CODES:
+            url = 'http://tools.ietf.org/html/rfc4918#section-11.%d' % (WEBDAV_STATUS_CODES.index(code) + 1)
         elif code in HTTP_STATUS_CODES:
             url = 'http://www.w3.org/Protocols/rfc2616/rfc2616-sec10.html' \
                   '#sec10.' + ('%d.%d' % (code // 100, 1 + code % 100))
@@ -619,15 +623,15 @@ class HTTPDomain(Domain):
             text = contnode.rawsource
             role = self.roles.get(typ)
             if role is None:
-                return nodes.emphasis(text, text)
+                return None
             resnode = role.result_nodes(env.get_doctree(fromdocname),
                                         env, node, None)[0][0]
             if isinstance(resnode, addnodes.pending_xref):
                 text = node[0][0]
                 reporter = env.get_doctree(fromdocname).reporter
-                reporter.error('Cannot resolve reference to %r' % text,
-                               line=node.line)
-                return nodes.problematic(text, text)
+                reporter.warning('Cannot resolve reference to %r' % text,
+                                 line=node.line)
+                return None
             return resnode
         else:
             anchor = http_resource_anchor(typ, target)