+++ lassodomain/CHANGES.rst
@@ -1,3 +1,14 @@
+0.3 (2015-01)
+-------------
+
+*  Fixes for Python 3.
+*  Added linking of parameter and return type names using the ``:type:`` role
+   and trait names using the ``:trait:`` role.
+*  New ``SingleGroupedField`` field type for grouped fields without an argument.
+*  Made reference matching case-insensitive.
+*  Added support for the ``:any:`` role.
+*  Added basic support for parallel builds.
+
 0.2 (2013-12)
 -------------